add base test classes for controller tests
