+++ ManagedCode.Storage.IntegrationTests/Tests/Azure/AzureDownloadControllerTests.cs
@@ -1,51 +1,10 @@
-﻿using System.Net;
-using FluentAssertions;
-using ManagedCode.Storage.Core.Models;
-using ManagedCode.Storage.IntegrationTests.Constants;
-using ManagedCode.Storage.IntegrationTests.Helpers;
-using Xunit;
+﻿using ManagedCode.Storage.IntegrationTests.Constants;
 
 namespace ManagedCode.Storage.IntegrationTests.Tests.Azure;
 
-public class AzureDownloadControllerTests : BaseControllerTests
+public class AzureDownloadControllerTests : BaseDownloadControllerTests
 {
-    public AzureDownloadControllerTests(StorageTestApplication testApplication) : base(testApplication)
+    public AzureDownloadControllerTests(StorageTestApplication testApplication) : base(testApplication, ApiEndpoints.Azure)
     {
-    }
-
-    [Fact]
-    public async Task DownloadFile_WhenFileExists_SaveToTempStorage_ReturnSuccess()
-    {
-        // Arrange
-        var storageClient = GetStorageClient();
-        var contentName = "file";
-
-        await using var localFile = LocalFile.FromRandomNameWithExtension(".txt");
-        FileHelper.GenerateLocalFile(localFile, 1);
-        var fileCRC = Crc32Helper.Calculate(await localFile.ReadAllBytesAsync());
-        var uploadFileBlob = await storageClient.UploadFile(localFile.FileStream, ApiEndpoints.Azure.UploadFile, contentName);
-        
-        // Act
-        var downloadedFileResult = await storageClient.DownloadFile(uploadFileBlob.Value.FullName, ApiEndpoints.Azure.DownloadFile);
-
-        // Assert
-        downloadedFileResult.IsSuccess.Should().BeTrue();
-        downloadedFileResult.Value.Should().NotBeNull();
-        var downloadedFileCRC = Crc32Helper.Calculate(await downloadedFileResult.Value.ReadAllBytesAsync());
-        downloadedFileCRC.Should().Be(fileCRC);
-    }
-
-    [Fact]
-    public async Task DownloadFile_WhenFileDoNotExist_ReturnFail()
-    {
-        // Arrange
-        var storageClient = GetStorageClient();;
-        
-        // Act
-        var downloadedFileResult = await storageClient.DownloadFile(Guid.NewGuid().ToString(), ApiEndpoints.Azure.DownloadFile);
-
-        // Assert
-        downloadedFileResult.IsFailed.Should().BeTrue();
-        downloadedFileResult.GetError().Value.ErrorCode.Should().Be(HttpStatusCode.InternalServerError.ToString());
     }
 }
